2057 Rename tooolkit to componentkit
