+++ sdks/backend/embedded/ai-componentkit/build.gradle.kts
