+++ settings.gradle.kts
@@ -508,8 +508,8 @@ include("server:libs:modules:task-dispatchers:terminate")
 include("server:libs:test:test-support")
 include("server:libs:test:test-int-support")
 
-include("sdks:backend:embedded:ai-toolkit")
-include("sdks:backend:embedded:ai-toolkit-spring")
+include("sdks:backend:embedded:ai-componentkit")
+include("sdks:backend:embedded:ai-componentkit-spring")
 include("sdks:backend:embedded:integration-api")
 include("sdks:backend:java:component-api")
 include("sdks:backend:java:component-test")
